MAKER-556: Handle multiple Serial.end() calls safely
Due to the use of pthreads in the Serial class, calling
Serial.end() twice consecutively causes the sketch to hang
(because pthread_join is waiting for a thread that has
already been joined). Use a bool to track wether Serial
is enabled, and don't perform init/exit tasks if Serial is
already enabled/disabled.